Fix issue with sliceAddress

+++ smartcontract/contracts/KYCVerification.sol
@@ -4,6 +4,7 @@ pragma solidity ^0.8.17;
 import "@openzeppelin/contracts/access/Ownable.sol";
 
 contract KYCVerification is Ownable {
+
     struct Signature {
         uint8 v;
         bytes32 r;
@@ -35,12 +36,12 @@ contract KYCVerification is Ownable {
 
     function sliceAddress(
         bytes calldata b,
-        uint offset
+        uint256 offset
     ) private pure returns (address) {
         bytes32 out;
-        for (uint i = 0; i < 20; i++) {
+        for (uint256 i = 0; i < 20; i++) {
             out |= bytes32(b[offset + i] & 0xFF) >> ((i + 12) * 8);
         }
-        return address(uint160(uint(out)));
+        return address(uint160(uint256(out)));
     }
 }
